splits out ProjectList into subcomponents
